♻️ JE-103-refactor: 파일명 수정
